[CMPT-5391] rename pipelines→pipeline in cmd/**, internal/drapi/**, and docs
- Rename cmd/pipelines/ → cmd/pipeline/ (Use: "pipeline", alias "pipelines")
- Rename internal/drapi/pipelines/ → internal/drapi/pipeline/ (API URLs unchanged)
- Change feature gate annotation to singular "pipeline" (env: DATAROBOT_CLI_FEATURE_PIPELINE)
- Add url.PathEscape for pipelineID in scope.go; add failing+passing path-escape tests
- Add unit tests for time.go (7 cases + struct unmarshal)
- Update docs to reference DATAROBOT_CLI_FEATURE_PIPELINE
