+++ cmd/pipeline/cmd_test.go
@@ -12,27 +12,34 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package pipelines
+package pipeline
 
 import (
+	"testing"
+
 	"github.com/datarobot/cli/internal/features"
-	"github.com/spf13/cobra"
+	"github.com/stretchr/testify/assert"
 )
 
-func Cmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:     "pipeline",
-		Aliases: []string{"pipelines"},
-		GroupID: "core",
-		Short:   "Pipelines API management commands",
-		Long: `Manage AI/ML pipelines orchestrated by Covalent.
+func TestCmd_BasicMetadata(t *testing.T) {
+	cmd := Cmd()
 
-Create, list, inspect, and update pipelines registered with the
-DataRobot pipelines service. Sub-commands are also available for managing
-input payloads, runs, and recurring schedules.`,
-	}
+	assert.Equal(t, "pipeline", cmd.Use)
+	assert.Equal(t, "core", cmd.GroupID)
+	assert.NotEmpty(t, cmd.Short)
+	assert.NotEmpty(t, cmd.Long)
+}
 
-	features.SetGate(cmd, "pipelines")
+func TestCmd_FeatureGate(t *testing.T) {
+	cmd := Cmd()
 
-	return cmd
+	gate, ok := cmd.Annotations[features.AnnotationKey]
+	assert.True(t, ok, "expected feature-gate annotation to be set")
+	assert.Equal(t, "pipeline", gate)
+}
+
+func TestCmd_NoSubcommandsYet(t *testing.T) {
+	cmd := Cmd()
+
+	assert.Empty(t, cmd.Commands(), "base branch registers no subcommands")
 }
